Moved folders to namespace style, modified setup.py and noxfile.py to adjust for that
