+++ tests/test_unit.py
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-import utils
+from google.cloud.sql import utils
 
 
 def test_generate_keys_not_return_none():
